% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights-bg.R
\encoding{utf-8}
\name{BGWeights}
\alias{BGWeights}
\title{Bates-Granger model weights}
\usage{
BGWeights(object, ..., data, seed = NULL, force.update = FALSE)
}
\arguments{
\item{object, \dots}{two or more fitted \code{\link{glm}} objects, or a
\code{list} of such, or an \code{\link[=model.avg]{"averaging"}} object.}

\item{data}{a data frame in which to look for variables for use with \link[=predict]{prediction}.}

\item{seed}{optionally, the random seed, see \code{\link{set.seed}} .}

\item{force.update}{if \code{TRUE}, the much less efficient method of updating \code{glm} function will be 
   used rather than directly \emph{via} \code{\link{glm.fit}}. This only applies to \code{glm}s, in 
case of other model types \code{update} is always used.}
}
\value{
The function returns a numeric vector of model weights.
}
\description{
Computes empirical weights based on out of sample forecast variances, following Bates and Granger (1969).
}
\details{
[TODO] Take a series of (pseudo) out of sample forecasts and forecast errors
Compute forecast variance (square of RMSE)
Invert
Normalize by sum across all models
}
\examples{
#[TODO example]
}
\author{
Carsten Dormann, Kamil Barto\enc{ń}{n}
}
\seealso{
\code{\link{Weights}}

Other model.weights: \code{\link{cos2Weights}},
  \code{\link{jackknifeWeights}},
  \code{\link{stackingWeights}}
}
\keyword{models}

