\name{Cement}
\alias{Cement}
\encoding{utf-8}
\docType{data}
\title{Cement hardening data}
\description{
Cement hardening data from Woods et al (1932).
}

\usage{
Cement
}

\format{
\code{Cement} is a data frame with 5 variables. \var{x1}-\var{x4} are four predictor
variables expressed as a percentage of weight.

\describe{
   \item{y}{calories of heat evolved per gram of cement after 180 days of
	hardening}  
   \item{X1}{calcium aluminate}
   \item{X2}{tricalcium silicate}
   \item{X3}{tetracalcium alumino ferrite}
   \item{X4}{dicalcium silicate.}

}}

\source{
Woods H., Steinour H.H., Starke H.R. (1932) Effect of composition of Portland
cement on heat evolved during hardening. \emph{Industrial & Engineering
Chemistry} 24, 1207--1214.
}

\references{
Burnham, K. P. and Anderson, D. R. 2002 \emph{Model selection and multimodel
inference: a practical information-theoretic approach}. 2nd ed. New York,
Springer-Verlag.
}

\keyword{datasets}
