\name{GPA}
\alias{GPA}
\encoding{utf-8}
\docType{data}
\title{Grade Point Average data}
\description{
First-year college Grade Point Average (\acronym{GPA}) from Graybill and Iyer (1994).
}

\usage{
GPA
}

\format{
\code{GPA} is a data frame with 5 variables. \var{y} is the first-year college
Grade Point Average (\acronym{GPA}) and \var{x1}-\var{x4} are four predictor variables
from standardized tests (\acronym{SAT}) administered before matriculation.


\describe{
   \item{y}{\acronym{GPA}}
   \item{x1}{math score on the \acronym{SAT}}
   \item{x2}{verbal score on the \acronym{SAT}}
   \item{x3}{high school math}
   \item{x4}{high school English}
}}

\source{
Graybill, F.A. and Iyer, H.K. (1994). \emph{Regression analysis: concepts and
applications}. Duxbury Press, Belmont, CA.
}

\references{
Burnham, K. P. and Anderson, D. R. 2002 \emph{Model selection and multimodel
inference: a practical information-theoretic approach}. 2nd ed. New York,
Springer-Verlag.
}

\keyword{datasets}
