\name{glm.link}
\alias{glm.link}
\encoding{utf-8}
\title{
 Get link from a GLM-type object
}
\description{
 Extract or build link from a fitted model object.
}
\usage{
glm.link(x)
}
\arguments{
  \item{x}{a fitted model object.}
}

\value{
A object of class \code{\link[=make.link]{"link-glm"}}. 
}

\author{Kamil Barto\enc{ń}{n}}

\seealso{\code{\link{make.link}}, \code{\link{power}}, \code{\link{glm}}, \code{\link{family}}.}

\keyword{manip}

