% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weights-jackknife.R
\encoding{utf-8}
\name{jackknifeWeights}
\alias{jackknifeWeights}
\title{Jackknifed model weights}
\usage{
jackknifeWeights(object, ..., data, method = c("RMSE", "likelihood"),
  weights = rep(1, n), optim.method = "BFGS", maxit = 1000, seed = NULL,
  optim.args = list(), force.update = FALSE)
}
\arguments{
\item{object, \dots}{two or more fitted \code{\link{glm}} objects, or a
\code{list} of such, or an \code{\link[=model.avg]{"averaging"}} object.}

\item{data}{a data frame in which to look for variables for use with
\link[=predict]{prediction}.}

\item{method}{the function to minimize. Either \code{"RMSE"} or
\code{"likelihood"} (yet only the former is currently implemented).}

\item{weights}{prior model weights.}

\item{optim.method}{optional, optimisation method, passed to \code{\link{optim}}.}

\item{maxit}{optional, the maximum number of iterations, passed to
\code{\link{optim}}.}

\item{seed}{optionally, the random seed, see \code{\link{set.seed}}.}

\item{optim.args}{optional list of other arguments passed to \code{\link{optim}}.}

\item{force.update}{for \code{glm}, the \code{glm.fit} function is used for
fitting models to the train data, which is much more efficient. Set to
\code{TRUE} to use \code{update} instead.}
}
\value{
The function returns a numeric vector of model weights.
}
\description{
Computes model weights optimized for jackknifed model fits.
}
\details{
Model weights are chosen (using \code{\link{optim}}) to minimise RMSE of
    the prediction for data point \var{i}, of a model fitted omitting that
    data point \var{i}. The jackknife procedure is therefore run for all
    provided models and for all data points.
}
\examples{
fm <- glm(Prop ~ mortality * dose, binomial(), Beetle, na.action = na.fail)

fits <- lapply(dredge(fm, eval = FALSE), eval)

amJk <- amAICc <- model.avg(fits)
Weights(amJk) <- jackknifeWeights(fits, data = Beetle)

coef(amJk)
coef(amAICc)

}
\author{
Carsten Dormann, Kamil Barto\enc{ń}{n}
}
\references{
Hansen, B. E. & Racine, J. S. (2012) Jackknife model averaging.
    \emph{Journal of Econometrics}, 979: 38–46
}
\seealso{
\code{\link{Weights}}

Other model.weights: \code{\link{BGWeights}},
  \code{\link{cos2Weights}}, \code{\link{stackingWeights}}
}
\keyword{models}

